+++ typescript/infra/config/docker.ts
@@ -20,6 +20,7 @@ export const DockerImageRepos = Object.fromEntries(
 interface AgentDockerTags {
   relayer: string;
   relayerRC: string;
+  relayerFastPath: string;
   validator: string;
   validatorRC: string;
   scraper: string;
@@ -37,8 +38,9 @@ interface MainnetDockerTags extends BaseDockerTags {
 
 export const mainnetDockerTags: MainnetDockerTags = {
   // rust agents
-  relayer: 'f6b6d7a-20260323-123723',
-  relayerRC: 'f6b6d7a-20260323-123723',
+  relayer: '6bbc7b5-20260326-002348',
+  relayerRC: '6bbc7b5-20260326-002348',
+  relayerFastPath: '3754d4f-20260326-130321',
   validator: '910e8e8-20260318-204227',
   validatorRC: '910e8e8-20260318-204227',
   scraper: '910e8e8-20260318-204227',
@@ -52,8 +54,9 @@ export const mainnetDockerTags: MainnetDockerTags = {
 
 export const testnetDockerTags: BaseDockerTags = {
   // rust agents
-  relayer: '910e8e8-20260318-204227',
-  relayerRC: '910e8e8-20260318-204227',
+  relayer: '6bbc7b5-20260326-002348',
+  relayerRC: '6bbc7b5-20260326-002348',
+  relayerFastPath: '3754d4f-20260326-130321',
   validator: '910e8e8-20260318-204227',
   validatorRC: '910e8e8-20260318-204227',
   scraper: '910e8e8-20260318-204227',